// EstimativaGrid_PCF/EstimativaGrid/models/enums.ts
/**
 * Enums for the PCF Estimativas control
 */

export enum ActivityType {
    Development = 'Development',
    Process = 'Process',
    Support = 'Support'
}

export enum Complexity {
    VeryLow = 'Very Low',
    Low = 'Low',
    Medium = 'Medium',
    High = 'High',
    VeryHigh = 'Very High'
}